+++ policyengine-client/src/countries/uk/components/timeTravel.jsx
@@ -6,53 +6,22 @@ import { useContext } from 'react';
 import { CountryContext } from '../../country';
 
 export default function TimeTravel(props) {
-    let [reformDate, setReformDate] = React.useState(moment());
+    let [date, setDate] = React.useState(moment());
     const country = useContext(CountryContext);
     let [isLoading, setIsLoading] = React.useState(false);
     return (
         <>
-            <h6 style={{marginTop: 20}}>Baseline snapshot</h6>
-            <p>PolicyEngine will use tax-benefit policy as of the date set below for the baseline simulation</p>
-            <DatePicker 
-                allowClear={false}
-                defaultValue={moment(country.policy.policy_date.value, "YYYYMMDD")} 
-                disabledDate={date => date < moment("2021-01-01") | date > moment("2021-12-31")}
-                onChange={(_, dateString) => {
-                    const url = `${country.apiURL}/parameters?policy_date=${dateString}`;
-                    fetch(url)
-                    .then((res) => {
-                        if (res.ok) {
-                            return res.json();
-                        } else {
-                            throw res;
-                        }
-                    }).then((policy) => {
-                        let previous_policy = country.policy;
-                        const dateInt = +(dateString.replace("-", "").replace("-", ""));
-                        for(let key in previous_policy) {
-                            previous_policy[key].defaultValue = policy[key].value;
-                        }
-                        previous_policy.policy_date.value = dateInt;
-                        previous_policy.policy_date.defaultValue = +(moment().format("YYYYMMDD"));
-                        country.updateEntirePolicy(previous_policy);
-                    }).catch(e => {
-                        message.error("Couldn't time travel - something went wrong." + e.toString());
-                        setIsLoading(false);
-                    });
-                }}
-            />
-            <div style={{paddingBottom: 20}} />
-            <h6 style={{marginTop: 20}}>Reform snapshot</h6>
-            <p>Select a date below to set all parameters to their legislative value as of that date</p>
+            <h6 style={{marginTop: 20}}>Snapshot</h6>
+            <p>Select a date below to set all parameters to their legislative value as of that date.</p>
             <DatePicker
                 allowClear={false}
-                defaultValue={reformDate}
-                onChange={date => setReformDate(date)}
-                disabledDate={date => date < moment("2021-01-01") | date > moment("2021-12-31")}
+                defaultValue={date}
+                onChange={date => setDate(date)}
+                disabledDate={date => date < moment("2021-01-01") | date > moment("2027-12-31")}
             />
             <Button 
                 onClick={() => {
-                    const dateString = reformDate.format("YYYY-MM-DD");
+                    const dateString = date.format("YYYY-MM-DD");
                     const url = `${country.apiURL}/parameters?policy_date=${dateString}`;
                     setIsLoading(true);
                     fetch(url)
@@ -65,14 +34,12 @@ export default function TimeTravel(props) {
                         }).then((policy) => {
                             let previous_policy = country.policy;
                             for(let key in policy) {
-                                if(key !== "baseline_policy_date") {
-                                    previous_policy[key].value = policy[key].value;
-                                }
+                                previous_policy[key][country.editingReform ? "value" : "baselineValue"] = policy[key].value;
                             }
                             country.updateEntirePolicy(previous_policy);
                             setIsLoading(false);
                         }).catch(e => {
-                            message.error("Couldn't time travel - something went wrong." + e.toString());
+                            message.error("Couldn't time travel - something went wrong.");
                             setIsLoading(false);
                         });
                 }}
@@ -80,6 +47,17 @@ export default function TimeTravel(props) {
             >Set policy</Button>
             {isLoading && <Spinner />}
             <div style={{paddingBottom: 20}} />
+            <h6 style={{marginTop: 20}}>Reset</h6>
+            <p>Reset all reform parameters to their values in the baseline.</p>
+            <Button onClick={
+                () => {
+                    let previous_policy = country.policy;
+                    for(let key in country.policy) {
+                        previous_policy[key].value = country.policy[key].baselineValue;
+                    }
+                    country.updateEntirePolicy(previous_policy);
+                }
+            }>Reset reform to baseline</Button>
         </>
   );
 }