+++ CHANGELOG.md
@@ -7,6 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 This repo consists of two packages - the React client and the Python server. A change to either repo should trigger an update in the versions for both to ensure a consistent changelog in this repo.
 
+## [1.25.1] - 2022-03-28 13:15:19
+
+### Fixed
+
+- Deployment timeout increased.
+
+## [1.25.0] - 2022-03-28 10:17:04
+
+### Added
+
+- Snapshot and reset buttons to the US app.
+
+### Changed
+
+- Baseline policy is now editable.
+
+### Fixed
+
+- HTTPS redirect sped up using HTML meta tags.
+
 ## [1.24.2] - 2022-03-24 21:40:50
 
 ### Fixed
@@ -547,6 +567,8 @@ This repo consists of two packages - the React client and the Python server. A c
 
 
 
+[1.25.1]: https://github.com/PolicyEngine/policyengine/compare/1.25.0...1.25.1
+[1.25.0]: https://github.com/PolicyEngine/policyengine/compare/1.24.2...1.25.0
 [1.24.2]: https://github.com/PolicyEngine/policyengine/compare/1.24.1...1.24.2
 [1.24.1]: https://github.com/PolicyEngine/policyengine/compare/1.24.0...1.24.1
 [1.24.0]: https://github.com/PolicyEngine/policyengine/compare/1.23.5...1.24.0